backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Cache
redis[hiredis]==5.0.1
hiredis==2.3.2

# Pydantic & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
cryptography==42.0.0

# LangChain & AI
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-community==0.0.17
langgraph==0.0.20
langsmith>=0.0.83,<0.1

# Vector Database
pymilvus==2.3.5
sentence-transformers==2.3.1

# Task Queue
celery[redis]==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Payment
python-alipay-sdk==3.4.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
